Now downloads the final power hour in thread
Threaded downloads now work for the final power hour export. Learned the
hard lesson that if you do anything with a thread in QT, the only way to
communicate back to the UI thread is through slots. You cannot even have
a reference to an object that is on the UI thread. Worker gets moved to
new QThread and calls out to the media downloading service, which now
has a bunch of callbacks to support the cool progress bars.