Simplifies export thread
Removes the worker class and the moveToThread call. I think these
weren't working together correctly and were causing some weird race
condition. Now there is just a QThread class and the code to do the
export is executed in its #run. I read somewhere that signals weren't
supposed to work correctly out of this but they seem to be functioning
fine.