Adds ability to migrate database
This commit adds handling for migrations located in sql files in the
db/migrations directory. It compares the name of the file, which is an
integer.sql, to the highest level present in the migrations table. If it
is greater than that level, it will execute the statements in the
migration file.

If any of the statements fail, it will rollback the changes and abort
loading the app.