Add first power hour in migration
This makes sure that there is a power hour present the first time
someone opens the application. This is important because it probably
wouldn't be immediately obvious how to create a new one. I decided to do
it in the migration #1's SQL so as to not have to write another,
parallel migration logic in python code.