Fixes insert row above with persistence
Adds a method that increments all of the positions above the row where
you are going to insert, so that the positions stay consistent.
Refactors insertion of a new power hour track and replaces the insertRow
function with it, as insertRow does not really work because the
power_hour_id and position must be set according to constraints.