Adds add track to end option to tracklist menu
Can now add a track to the end of the power hour with the context menu.
This gets rid of the add track button on the main window. This commit
also checks to make sure there are tracks before showing the insert
above, below, and delete options are shown in the context menu.