Removes downloading from media processors
I pulled the media downloading out of the media processor class, because
with local files the downloader that I need could be different for each
track. Next step will be to remove the relevant code from the
MediaProcessor and CreatePowerHourService classes. The media.py file has also been rearragned to be in chronological order of
calls from the gui: find_track -> CreatePowerHourService.