Refactors main_window, config, and models
Many responsibilities were removed from main window. The models for
power hour list and track list are now build from outside and dependency
injected in. Creation of the export dialog and thread moved to separate
functions in the ui/exporting module, along with those classes. Audio
and video formats in strings moved to config variables so those are
easier to switch out if we want. Track length moved to config.