Validates that start time is less than track length
To do the validation, I added a start time validator class that right
now just checks for that. To wire it up with the main window, I created
a dispatcher class that will just a signal, and dependency injected that
into the main window and into the delgate.

Within the tests, I started to clean up the messy tracklist interaction
code, and put it in the TracklistTestModel class, which is inspired by
the Page objects you sometimes see in Rails tests.