Stores last export directory for better ux
Modifies the ExportLocator to store the last directory in the persistent
settings so that when you go to export again you are already in your
preferred folder.

This commit also makes the config.OS variable which makes access to the
name of the operating system consitent and a little faster (not that it
mattered anyways).