Adds json export of power hours
This commit adds an export current tracklist option to the file menu,
which creates an export of the power hour currently selected. To do
this, I used the attr.asdict method on track objects.

This commit also includes many refactors to test files to support using
pytest for features in the future. There is an issue using pytest-qt
with my bootstrapped app and main_window, so I am avoiding using any of
their fixtures in the feature tests.