Fixes bug with track deletion
Track position updating query did not properly account for the order of
the rows in the update. It was possible to violate the unique constraint
on tracks.position when the query decremented the position, if the rows
weren't returned in ascending order by position. Since SQLITE doesn't
have an order by clause for updates, I had to add the extra query.

Adds test to verify bug in last commit fixed