Saves last import location in settings
PHC now saves the last dir you used to import a tracklist from in the
settings. In doing this, a pretty big refactor emerged for tests and for
saving things in settings. I wrote a decorator to store function results
in settings, which should make it a little easier to do so.