Fixes bugs with deleting power hour and threading
Previously, if a user would delete the last power hour in the list, it
would not select another power hour for them. We now check to see if the
user deleted the last in the list, and then select the previous entry.

Fixed a bug which never arose on windows, but the creation thread
finished signal was linked to the deletelater slot. The slot now would
fire before the execution of the thread was finished, and that would
cause qt to throw an error. We now no longer connect to the delete later
slot, and everything still works as it should.