Gets basic add new power hour and export working with sqlalchemy
This commit adds the necessary methods to the sqlalchemy based
TracklistModel to get it working with adding new power hours and just
exporting a simple power hour. Much more will have to be done to support
the other methods available in the context menu, but this is a good
start for now.